Switch to using 16 bit addresses for memory

+++ src/goboye/cpu/opcode_handlers_test.go
@@ -22,8 +22,8 @@ func TestNopHandler(t *testing.T) {
 	o := p.readNextInstruction()
 	o.handler(o, p)
 
-	assert.Equal(t, 1, p.programCounter)
-	assert.Equal(t, 0, p.stackPointer)
+	assert.Equal(t, uint16(1), p.programCounter)
+	assert.Equal(t, uint16(0), p.stackPointer)
 }
 
 func TestLoad16BitImmediate(t *testing.T) {
@@ -31,8 +31,8 @@ func TestLoad16BitImmediate(t *testing.T) {
 	o := p.readNextInstruction()
 	o.handler(o, p)
 
-	assert.Equal(t, 3, p.programCounter)
-	assert.Equal(t, 0, p.stackPointer)
+	assert.Equal(t, uint16(3), p.programCounter)
+	assert.Equal(t, uint16(0), p.stackPointer)
 	assert.Equal(t, uint16(0x1234), p.registers.bc)
 }
 
@@ -44,8 +44,8 @@ func TestSaveAtoBCAddr(t *testing.T) {
 	p.registers.bc = 0x0004
 	o.handler(o, p)
 
-	assert.Equal(t, 1, p.programCounter)
-	assert.Equal(t, 0, p.stackPointer)
+	assert.Equal(t, uint16(1), p.programCounter)
+	assert.Equal(t, uint16(0), p.stackPointer)
 	assert.Equal(t, uint8(0x56), p.memory.ReadByte(0x0400))
 	assert.Equal(t, uint8(0x56), p.registers.getRegister(A))
 }
